@@ -38,24 +37,31 @@ base {
     archivesName = mod_id
 }
 
-// Mojang ships Java 17 to end users in 1.20.1, so mods should target Java 17.
-java.toolchain.languageVersion = JavaLanguageVersion.of(17)
+// Mojang ships Java 21 to end users in 1.21.10, so mods should target Java 21.
+java.toolchain.languageVersion = JavaLanguageVersion.of(21)
 
-legacyForge {
-    // Specify the version of MinecraftForge to use.
-    version = project.minecraft_version + '-' + project.forge_version
+neoForge {
+    // Specify the version of NeoForge to use.
+    version = project.neo_version
 
     parchment {
         mappingsVersion = project.parchment_mappings_version
         minecraftVersion = project.parchment_minecraft_version
     }
 
+    // This line is optional. Access Transformers are automatically detected
+     accessTransformers = project.files('src/main/templates/META-INF/accesstransformer.cfg')
+
     // Default run configurations.
+    // These can be tweaked, removed, or duplicated as needed.
     runs {
         client {
             client()
-            systemProperty 'forge.enabledGameTestNamespaces', project.mod_id
+
+            // Comma-separated list of namespaces to load gametests from. Empty = all namespaces.
+            systemProperty 'neoforge.enabledGameTestNamespaces', project.mod_id
         }
+
         clientAuth{
             devLogin = true
             client()
