@@ -1,12 +1,15 @@
 package top.r3944realms.lib39.example.content.capability;
 
-import net.minecraft.world.entity.Entity;
-import net.minecraftforge.common.capabilities.Capability;
-import net.minecraftforge.common.capabilities.CapabilityManager;
-import net.minecraftforge.common.capabilities.CapabilityToken;
-import net.minecraftforge.common.capabilities.RegisterCapabilitiesEvent;
-import net.minecraftforge.event.AttachCapabilitiesEvent;
-import org.jetbrains.annotations.NotNull;
+import net.minecraft.resources.ResourceLocation;
+
+import net.minecraft.world.entity.LivingEntity;
+import net.minecraft.world.entity.vehicle.Boat;
+import net.minecraft.world.entity.vehicle.Minecart;
+import net.neoforged.neoforge.capabilities.EntityCapability;
+import net.neoforged.neoforge.capabilities.RegisterCapabilitiesEvent;
+import top.r3944realms.lib39.Lib39;
+import top.r3944realms.lib39.example.core.register.ExLib39Attachments;
+import top.r3944realms.lib39.util.capability.EntityCapabilityHelper;
 
 /**
  * The type Ex capability handler.
@@ -15,26 +18,29 @@ public class ExCapabilityHandler {
     /**
      * The constant TEST_CAP.
      */
-    public static final Capability<AbstractedTestSyncData> TEST_CAP = CapabilityManager.get(new CapabilityToken<>() {});
+    public static final EntityCapability<AbstractedTestSyncData, Void> TEST_CAP = EntityCapability.createVoid(
+            ResourceLocation.fromNamespaceAndPath(Lib39.MOD_ID, "test_data"),
+            AbstractedTestSyncData.class
+    );
 
-    /**
-     * Register capability.
-     *
-     * @param event the event
-     */
-    public static void registerCapability(@NotNull RegisterCapabilitiesEvent event) {
-        event.register(AbstractedTestSyncData.class);
-    }
-
-    /**
-     * Attach capability.
-     *
-     * @param event the event
-     */
-    public static void attachCapability(@NotNull AttachCapabilitiesEvent<?> event) {
-        Object object = event.getObject();
-        if(object instanceof Entity entity ) {
-            event.addCapability(TestSyncCapProvider.TEST_SYNC_REL, new TestSyncCapProvider(entity));
-        }
+    public static void registerCapabilities(RegisterCapabilitiesEvent event) {
+        EntityCapabilityHelper.registerForEntityClass(
+                event,
+                TEST_CAP,
+                (entity, context) -> entity.getData(ExLib39Attachments.TEST_DATA_ATTACHMENT),
+                LivingEntity.class
+        );
+        EntityCapabilityHelper.registerForEntityClass(
+                event,
+                TEST_CAP,
+                (entity, context) -> entity.getData(ExLib39Attachments.TEST_DATA_ATTACHMENT),
+                Boat.class
+        );
+        EntityCapabilityHelper.registerForEntityClass(
+                event,
+                TEST_CAP,
+                (entity, context) -> entity.getData(ExLib39Attachments.TEST_DATA_ATTACHMENT),
+                Minecart.class
+        );
     }
 }
