@@ -1,13 +1,17 @@
 package top.r3944realms.lib39.example.core.register;
 
+
+import net.minecraft.core.registries.Registries;
 import net.minecraft.world.item.Item;
-import net.minecraftforge.eventbus.api.IEventBus;
-import net.minecraftforge.registries.DeferredRegister;
-import net.minecraftforge.registries.ForgeRegistries;
-import net.minecraftforge.registries.RegistryObject;
+
+import net.neoforged.bus.api.IEventBus;
+import net.neoforged.neoforge.registries.DeferredHolder;
+import net.neoforged.neoforge.registries.DeferredRegister;
 import top.r3944realms.lib39.Lib39;
 import top.r3944realms.lib39.example.content.item.FabricItem;
 import top.r3944realms.lib39.example.content.item.NeoForgeItem;
+
+import java.util.Objects;
 
 /**
  * The type Ex lib 39 items.
@@ -16,27 +20,29 @@ public class ExLib39Items {
     /**
      * The constant ITEMS.
      */
-    public static final DeferredRegister<Item> ITEMS = DeferredRegister.create(ForgeRegistries.ITEMS, Lib39.MOD_ID);
+    public static final DeferredRegister<Item> ITEMS = DeferredRegister.create(Registries.ITEM, Lib39.MOD_ID);
     /**
      * The constant SUPER_LEAD_ROPE.
      */
-    public static final RegistryObject<Item> FABRIC = ITEMS.register(
+    public static final DeferredHolder<Item, Item> FABRIC = ITEMS.register(
             "fabric",
             () -> new FabricItem(
                     new Item.Properties()
                             .stacksTo(1)
                             .fireResistant()
+                            .setId(Objects.requireNonNull(ExLib39ItemResourceKeys.FABRIC.getResourceKey()))
             )
     );
     /**
      * The constant ETERNAL_POTATO.
      */
-    public static final RegistryObject<Item> NEOFORGE =
+    public static final DeferredHolder<Item, Item> NEOFORGE =
             ITEMS.register("neoforge",
                     () -> new NeoForgeItem(
                             new Item.Properties()
                                     .stacksTo(1)
                                     .fireResistant()
+                                    .setId(Objects.requireNonNull(ExLib39ItemResourceKeys.NEOFORGE.getResourceKey()))
                     ));
 
     /**
